move to database-util and remove DatabaseCompilation
